+++ hypers/_learning/__init__.py
@@ -1,0 +1,6 @@
+from hypers._learning._cluster import _data_cluster
+from hypers._learning._vca import _vca
+from hypers._learning._abundance import _ucls
+from hypers._learning._decomposition import _data_decomposition, _data_scree
+
+__all__ = ['_data_cluster', '_vca', '_ucls', '_data_decomposition', '_data_scree']